# docs/swagger.yaml
basePath: /
definitions:
...
  controller.createTenantReq:
    properties:
      name:
        type: string
      parent_tenant_id:
        type: string
    required:
    - name
    type: object
...
  controller.listResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/controller.tenantResp'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  controller.tenantResp:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      parent_tenant_id:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: Central Authentication Service for multi-tenant identity management.
  title: Guard CAS API
  version: "1.0"
paths:
...
  /api/v1/tenants:
    get:
      description: Lists tenants with optional filters and pagination
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: -1 any, 1 active, 0 inactive
        in: query
        name: active
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.listResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List tenants
      tags:
      - tenants
    post:
      consumes:
      - application/json
      description: Creates a new tenant
      parameters:
      - description: name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.createTenantReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.tenantResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create tenant
      tags:
      - tenants
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"